requirements.txt
imutils==0.5.3
Keras==2.0.8
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
matplotlib==3.1.2
numpy==1.18.1
opencv-contrib-python==3.4.2.16
Pillow==7.0.0
tensorflow==1.13.1
